+++ packages/gestalt/src/Layer.js
@@ -5,16 +5,31 @@ import { createPortal } from 'react-dom';
 type Props = {|
   children?: React.Node,
 |};
+type State = {|
+  mounted: boolean,
+|};
 
-export default class Layer extends React.Component<Props> {
+export default class Layer extends React.Component<Props, State> {
+  state = {
+    mounted: false,
+  };
+
   constructor(props: Props) {
     super(props);
-    this.el = document.createElement('div');
+    if (typeof document !== 'undefined' && document.createElement) {
+      this.el = document.createElement('div');
+    } else {
+      // eslint-disable-next-line no-console
+      console.warn(
+        'Using Layer without document present. Children will not be rendered.'
+      );
+    }
   }
 
   componentDidMount() {
-    if (document.body) {
+    if (typeof document !== 'undefined' && document.body) {
       document.body.appendChild(this.el);
+      this.setState({ mounted: true });
     }
   }
 
@@ -28,6 +43,6 @@ export default class Layer extends React.Component<Props> {
 
   render() {
     const { children } = this.props;
-    return createPortal(children, this.el);
+    return this.state.mounted && createPortal(children, this.el);
   }
 }