+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Internal: Reverts upgrade to `postcss-preset-env` due to a bug in how it interacts with CSS modules (#451)
 - Box: update docs to mention new `ref` functionality (#450)
+- Layer: Warn instead of erroring when server-rendering (#454)
 
 ### Patch